class-39-django-rest-framework/demos/python-demo-kanban/requirements.txt
appnope==0.1.0
autopep8==1.4.3
backcall==0.1.0
confusable-homoglyphs==3.2.0
decorator==4.4.0
Django==2.2
django-cors-headers==2.5.2
django-registration==3.0
djangorestframework==3.9.2
entrypoints==0.3
factory-boy==2.11.1
Faker==1.0.4
flake8==3.7.7
gunicorn==19.9.0
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2-binary==2.8.1
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.3.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
